Avoid using an intermediate buffer when possible.
Eliminate the RedisResponse type, just have the handler directly write to the stream.  This mostly allows the intermediate buffer to be removed, except when we get a value.  There we have to make a copy of the value in case the reference gets invalidated during the await point on the write.